+++ example/lib/main.dart
@@ -3,6 +3,8 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:seon_sdk_flutter_plugin/seon_sdk_flutter_plugin.dart';
+import 'package:seon_sdk_flutter_plugin/seon_sdk_geolocation_config.dart';
+import 'package:seon_sdk_flutter_plugin/seon_sdk_geolocation_config_builder.dart';
 
 void main() {
   runApp(const MyApp());
@@ -23,8 +25,8 @@ class _MyAppState extends State<MyApp> {
   void initState() {
     super.initState();
     try {
-      _seonSdkFlutterPlugin.setGeolocationEnabled(true);
-      _seonSdkFlutterPlugin.setGeolocationTimeout(500);
+      final geoConfig = SeonGeolocationConfigBuilder().withGeolocationEnabled(true).withLocationServiceTimeoutMs(5000).withPrefetchEnabled(true).withMaxLocationCacheAgeSec(10).build();
+      _seonSdkFlutterPlugin.setGeolocationConfig(geoConfig);
     } catch (e) {
       print('$e');
     }
@@ -42,7 +44,24 @@ class _MyAppState extends State<MyApp> {
     }
 
     if (!mounted) return;
+    print(fingerprint);
+    setState(() {
+      _fingerprint = fingerprint;
+    });
+  }
 
+    Future<void> stopBehaviourMonitoring() async {
+    String fingerprint;
+    try {
+      fingerprint = await _seonSdkFlutterPlugin
+              .stopBehaviourMonitoring("flutter-behaviour-demo-session-id-2222") ??
+          'Error getting fingerprint';
+    } catch (e) {
+      fingerprint = 'Failed to get fingerprint $e';
+    }
+
+    if (!mounted) return;
+    print(fingerprint);
     setState(() {
       _fingerprint = fingerprint;
     });
@@ -54,6 +73,20 @@ class _MyAppState extends State<MyApp> {
     ScaffoldMessenger.of(context).showSnackBar(
       const SnackBar(content: Text('Copied to clipboard')),
     );
+  }
+  Future<void> startBehaviourMonitoring() async {
+    String fingerprint = "";
+    print('Behaviour started');
+    try {
+      _seonSdkFlutterPlugin.startBehaviourMonitoring();
+    } catch (e) {
+      fingerprint = 'Failed to start behaviour $e';
+      if (mounted) {
+        setState(() {
+          _fingerprint = fingerprint;
+        });
+      }
+    }
   }
 
   @override
@@ -84,6 +117,14 @@ class _MyAppState extends State<MyApp> {
                 ElevatedButton(
                   onPressed: () => copyToClipboard(_fingerprint),
                   child: const Text('Copy to Clipboard'),
+                ),
+                ElevatedButton(
+                  onPressed: startBehaviourMonitoring,
+                  child: const Text('Start behaviour'),
+                ),
+                ElevatedButton(
+                  onPressed: stopBehaviourMonitoring,
+                  child: const Text('Stop behaviour'),
                 ),
               ],
             ),