+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.1.0
+- Introduced Geolocation config object
+- Introduced Behavioural monitoring support
+- Supports Android SDK version: `6.6.0`. For the changelog visit: https://github.com/seontechnologies/seon-android-sdk-public#660
+- Supports iOS SDK version: `5.5.1` For the changelog visit: https://github.com/seontechnologies/seon-ios-sdk-public#551
+- Added fixes and improvements
+## 1.0.1
+- Fixed error propagation
 ## 1.0.0
 - Initial release
 - Supports Android SDK version: `6.4.1`. For the changelog visit: https://github.com/seontechnologies/seon-android-sdk-public#641
 - Supports iOS SDK version: `5.3.0` For the changelog visit: https://github.com/seontechnologies/seon-ios-sdk-public#530
-## 1.0.1
-- Fixed error propagation